Shrank data types

+++ src/pf_astro.c
@@ -3,17 +3,17 @@
 #include "pf_moon.h"
 
 
-int gmt_year;
-int gmt_mon;
-int gmt_mday;
-int gmt_hour;
-int gmt_min;
-int gmt_sec;
+int8_t gmt_year;
+int8_t gmt_mon;
+int8_t gmt_mday;
+int16_t gmt_hour;
+int8_t gmt_min;
+int8_t gmt_sec;
 double gmt_decitime;
-int gmt_juldate;
+uint32_t gmt_juldate;
 
 
-double hms_to_decimal (int hours, int minutes, int seconds) {
+double hms_to_decimal (int8_t hours, int8_t minutes, int8_t seconds) {
 		double decimal =
 				((double)hours)
 				+ (((double)minutes) / PF_MINS_PER_HOUR)
@@ -22,8 +22,8 @@ double hms_to_decimal (int hours, int minutes, int seconds) {
 }
 
 
-int julian_date (int year, int month, int day) {
-		int a, y, m;
+uint32_t julian_date (int16_t year, int8_t month, int8_t day) {
+		int8_t a, y, m;
 		a = (14 - month) / 12;
 		y = year + 4800 - a;
 		m = month + (12 * a) - 3;
@@ -34,17 +34,17 @@ int julian_date (int year, int month, int day) {
 				- (y / 100)
 				+ (y / 400)
 				- 32045;*/
-		int jul = day
-				+ (int)(((153 * m) + 2) / 5)
+		uint32_t jul = day
+				+ (uint32_t)(((153 * m) + 2) / 5)
 				+ (365 * y)
-				+ (int)(y / 4)
+				+ (uint32_t)(y / 4)
 				- 32083;
 		return jul;
 }
 
 
-int moon_phase (int julday) {
-		return julday % PF_NUM_PHASES;
+int8_t moon_phase (uint32_t julday) {
+		return (int8_t)(julday % PF_NUM_PHASES);
 }
